Break cyclical dependency
Make sure all of the customize search functionality is not in the main bundle. Shaves off 200KB from the initial load
