+++ app/util/planParamUtil.js
@@ -3,8 +3,19 @@ import omitBy from 'lodash/omitBy';
 import moment from 'moment';
 // Localstorage data
 import { getCustomizedSettings } from '../store/localStorage';
-import { defaultSettings } from '../component/CustomizeSearch';
 import { otpToLocation } from './otpStrings';
+
+export const WALKBOARDCOST_DEFAULT = 600;
+
+export const defaultSettings = {
+  accessibilityOption: 0,
+  minTransferTime: 120,
+  walkBoardCost: WALKBOARDCOST_DEFAULT,
+  transferPenalty: 0,
+  walkReluctance: 2,
+  walkSpeed: 1.2,
+  ticketTypes: 'none',
+};
 
 function getIntermediatePlaces(intermediatePlaces) {
   if (!intermediatePlaces) {
